Move journal feature out of shared folder
